Se crea documento controller del componente navbar con toda la parte funcional, se adicionan el value en el profile y el controler para llamar el nombre del usuario logueado en profile, y se hacen las primeras pruebas del metodo put para modificar el perfil las cuales no son exitosas por ahora
